pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.0.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
]
